Feature back/ticket spring

+++ back-spring/src/main/java/ssafy/nfticket/dto/request/block/SimpleBlockHashTicketDto.java
@@ -7,12 +7,12 @@ import lombok.Data;
 public class SimpleBlockHashTicketDto {
 
     @JsonProperty(value = "ticket_id")
-    private Integer ticketId;
+    private Long ticketId;
 
     @JsonProperty(value = "block_hash")
     private String blockHash;
 
-    public SimpleBlockHashTicketDto(Integer ticketId, String blockHash) {
+    public SimpleBlockHashTicketDto(Long ticketId, String blockHash) {
         this.ticketId = ticketId;
         this.blockHash = blockHash;
     }
